cleaned up handling of standard comparisons for integer column types

+++ src/main/java/com/github/lwhite1/tablesaw/columns/ShortColumn.java
@@ -95,7 +95,7 @@ public class ShortColumn extends AbstractColumn implements ShortMapUtils {
     data.set(index, value);
   }
 
-  public RoaringBitmap isLessThan(short i) {
+  public RoaringBitmap isLessThan(int i) {
     return apply(ShortColumnUtils.isLessThan, i);
   }
